+++ packages/extension/src/components/App.tsx
@@ -7,7 +7,7 @@ import { LinkSelector } from "@/components/LinkSelector"
 import { OpenInTab } from "@/components/OpenInTab"
 import { PageActionRecorder } from "@/components/pageAction/PageActionRecorder"
 import { PageActionRunner } from "@/components/pageAction/PageActionRunner"
-import { PageActionContextProvider } from "@/hooks/pageAction/usePageActionContext"
+import { PageActionContextProvider } from "@/providers/pageActionContextProvider"
 import { SelectContextProvider } from "@/providers/SelectContextProvider"
 import { TabContextProvider } from "@/providers/TabContextProvider"
 import { Ipc, TabCommand } from "@/services/ipc"
